refactor(linter): move DiagnosticsReporters to oxlint
